Move a test-only file into ./test/utils
